+++ workspaces/features-demo/src/demo-pieces/navigation-repeats-endings.ts
@@ -10,7 +10,7 @@ export function createNavigationRepeatEndingDemo() {
         .setTimeSignature("3/4")
         .addNote(0, "A3", "4n")
         .addNote(0, "A3", "2n")
-        .addAnnotation("endRepeat", { repeatCount: 3 })
+        .addAnnotation("endRepeat", { playCount: 3 })
 
         .addMeasure()
         .addAnnotation("startRepeat")
@@ -19,7 +19,7 @@ export function createNavigationRepeatEndingDemo() {
         .endRow()
 
         .addMeasure()
-        .addAnnotation("ending", { endingPassages: [1, 2] })
+        .addAnnotation("ending", { playNumbers: [1, 2] })
         .addNote(0, "A3", "4n")
         .addNote(0, "A3", "2n")
 
@@ -30,7 +30,7 @@ export function createNavigationRepeatEndingDemo() {
         .endRow()
 
         .addMeasure()
-        .addAnnotation("ending", { endingPassages: 3 })
+        .addAnnotation("ending", { playNumbers: 3 })
         .addNote(0, "A3", "4n")
         .addNote(0, "A3", "2n")
 
@@ -45,13 +45,13 @@ export function createNavigationRepeatEndingDemo() {
         .addNote(0, "A3", "2n")
 
         .addMeasure()
-        .addAnnotation("ending", { endingPassages: 1 })
+        .addAnnotation("ending", { playNumbers: 1 })
         .addNote(0, "A3", "4n")
         .addNote(0, "A3", "2n")
         .addAnnotation("endRepeat")
 
         .addMeasure()
-        .addAnnotation("ending", { endingPassages: 2 })
+        .addAnnotation("ending", { playNumbers: 2 })
         .addNote(0, "A3", "2.")
 
         .getDocument();